requirements.txt
Flask
Flask-SQLAlchemy
pymysql
flask-marshmallow
flask-jwt-extended
sqlalchemy-utils
marshmallow-sqlalchemy
requests
